Add support for arm64
Depends on corresponding API change in user space.

* Adds support for arm64
* Many small fixes to make generic code compile for arm64
* Moves teesmc*.h from armv7 specific to include/arm_common

+++ generic/tee_supp_com.h
@@ -91,10 +91,10 @@ enum teec_rpc_result tee_supp_cmd(struct tee_targetop *op,
 		uint32_t id, void *data,
 		unsigned int datalen);
 
-int tee_supp_read(struct file *filp, char __user *buffer,
+ssize_t tee_supp_read(struct file *filp, char __user *buffer,
 			 size_t length, loff_t *offset);
 
-int tee_supp_write(struct file *filp, const char __user *buffer,
+ssize_t tee_supp_write(struct file *filp, const char __user *buffer,
 			  size_t length, loff_t *offset);
